Deduplicate lifecycle-events mixins
