+++ legacy-fabric-lifecycle-events-v1/1.8.9/src/main/resources/legacy-fabric-lifecycle-events-v1.mixins.json
@@ -3,14 +3,10 @@
   "package": "net.legacyfabric.fabric.mixin.event.lifecycle",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "MinecraftServerMixin",
-    "ServerChunkProviderMixin",
-    "ServerWorldMixin"
+    "versioned.MinecraftServerMixin"
   ],
   "client": [
-    "client.ClientChunkProviderMixin",
-    "client.ClientWorldMixin",
-    "client.MinecraftClientMixin"
+    "client.versioned.ClientChunkProviderMixin"
   ],
   "injectors": {
     "defaultRequire": 1
