+++ legacy-fabric-lifecycle-events-v1/common/src/main/java/net/legacyfabric/fabric/api/client/event/lifecycle/v1/ClientTickEvents.java
