+++ legacy-fabric-lifecycle-events-v1/common/src/main/java/net/legacyfabric/fabric/mixin/event/lifecycle/ServerChunkProviderMixin.java
