+++ legacy-fabric-lifecycle-events-v1/common/src/main/java/net/legacyfabric/fabric/mixin/event/lifecycle/client/ClientChunkProviderMixin.java
@@ -17,12 +17,10 @@
 
 package net.legacyfabric.fabric.mixin.event.lifecycle.client;
 
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
@@ -39,17 +37,8 @@ import net.legacyfabric.fabric.api.client.event.lifecycle.v1.ClientChunkEvents;
 @Environment(EnvType.CLIENT)
 @Mixin(ClientChunkProvider.class)
 public abstract class ClientChunkProviderMixin {
-	@Final
 	@Shadow
 	private World world;
-
-	@Shadow
-	public abstract Chunk getLoadedChunk(int par1, int par2);
-
-	@Inject(at = @At("RETURN"), method = "unloadChunk")
-	public void chunkUnload(int i, int j, CallbackInfo ci) {
-		ClientChunkEvents.CHUNK_UNLOAD.invoker().onChunkUnload((ClientWorld) this.world, this.getLoadedChunk(i, j));
-	}
 
 	@Inject(at = @At("RETURN"), method = "getOrGenerateChunk", locals = LocalCapture.CAPTURE_FAILEXCEPTION)
 	public void chunkLoad(int i, int j, CallbackInfoReturnable<Chunk> cir, Chunk chunk) {
