+++ legacy-fabric-lifecycle-events-v1/common/src/main/resources/fabric.mod.json
@@ -18,7 +18,9 @@
   "depends": {
     "minecraft": "${minecraft_version}"
   },
-  "mixins": [],
+  "mixins": [
+    "legacy-fabric-lifecycle-events-v1-common.mixins.json"
+  ],
   "description": "Events for the game's lifecycle.",
   "custom": {
     "modmenu": {
