+++ legacy-fabric-lifecycle-events-v1/common/src/main/resources/legacy-fabric-lifecycle-events-v1-common.mixins.json
@@ -3,10 +3,14 @@
   "package": "net.legacyfabric.fabric.mixin.event.lifecycle",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "versioned.MinecraftServerMixin"
+    "MinecraftServerMixin",
+    "ServerChunkProviderMixin",
+    "ServerWorldMixin"
   ],
   "client": [
-    "client.versioned.ClientChunkProviderMixin"
+    "client.ClientChunkProviderMixin",
+    "client.ClientWorldMixin",
+    "client.MinecraftClientMixin"
   ],
   "injectors": {
     "defaultRequire": 1
